+++ src/main/java/com/codeaddi/mrc_resources/model/http/ResourceUseDTO.java
@@ -1,6 +1,7 @@
 package com.codeaddi.mrc_resources.model.http;
 
 import com.codeaddi.mrc_resources.model.enums.EquipmentType;
+import com.codeaddi.mrc_resources.model.http.enums.ResourceStatus;
 import lombok.extern.jackson.Jacksonized;
 
 import java.time.LocalDateTime;
